Add Admin Day 1 Tourpoints
Builds off of previous work that has been done with Teacher
and student  tour points.
These tour points only appear for users that have an admin role.

Adjusts the Courses Controller to include whether the current
user is a teacher in the currently viewed course as part of the
values provided to the front end.

When a user is an admin, teacher, and student in a course, the
admin tour immediately appears. Teacher and student tours
appear when the user visits the course page which they are a
teacher or student in.

Splits the tours into their own separate files to shorten files.

Adds a Selenium test for initial launch of the tour.

Test Plan:
Make sure the Product Tours feature flag is turned on
As an admin, open up canvas.
The tour should appear, allowing you to walk through the
steps.

As a admin, student, and teacher, open up canvas.
The admin tour should appear (unless it's already appeared)
Visit the course page for the course you are enrolled in.
The student tour should appear.
Visit the course page for the course you are teaching.
The teacher tour should appear.

flag = product_tours

Refs UXS-58

Change-Id: I254822836d7884cd182cfcf67cddde53824afca5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/232654
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Product-Review: Laura Leavitt <lleavitt@instructure.com>
Reviewed-by: Augusto Callejas <acallejas@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>