Fix autosubmission due to non-time limit end_at dates
This change ensures that the following requirement is respected:
The tool should only prevent submission on the inner most layer and that
is the attempt timer. All other layers such as Until date, Course end
date need to be respected.

Test plan:
1. Create a quiz without a time limit and observe that it will
   autosubmit when the earliest of the following times is reached: quiz
availability end date and course availability end date.
2. Now add a time limit and check Disable Timer Autosubmission and take
   the same quiz. You should observe that the clock visible counts down
to 0 for the duration of the time limit, yet does not autosubmit when
the clock reaches zero, but does autosubmit when the end time as
specified in step 1 is reached.

Refs PFS-15660
flag=timer_without_autosubmission

Change-Id: I8ac898203070b79221afb168ef96e75d8f68fdab
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/235571
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Stephen Kacsmark <skacsmark@instructure.com>
Reviewed-by: Chad McGuire <cmcguire@instructure.com>
QA-Review: Petra Ashment <pashment@instructure.com>
Product-Review: Chad McGuire <cmcguire@instructure.com>