AGS: Fix for duplicate line items created
in `create_line_item!`, instead of preventing the Assignment from
creating the default LineItem, and having to create the
`Lti::ResourceLink` ourselves, it actually seemed easier to just let it
create the default line item and then update it with the attributes from
the params. Let me know if you think that's not OK or there's a better
way. I don't know if we were ever planning on changing that "default
line item" logic, but doing this avoids some weird dependencies (the
ResourceLink should be created with a resource_link_id equal to the
Assignment's lti_context_id, the line item should be created with a
resource_link, but the assignment should be created by passing in a line
item).

NOTE: In the case where a URL is passed-in that does not correspond to
an existing tool, the ContentTag created (by
external_tool_tag_attributes) has a url which does not correspond to any
tool, so ContextExternalTool.from_content_tag fails and we create a line
item but no ResourceLink. I didn't change this behavior, but I'm not
sure what the behavior should be.

ALSO NOTE: in the case where a line item is created with
submission_type=none, no ResourceLink is created. This behavior has not
changed.

flag=none
closes PLAT-5720

Test plan:
- Install the LTI 1.3 test tool and go to
  http://web.lti-13-test-tool.docker/ags/new
- Create a new line item: fill in credential ID (from your local test
  tool DB), context ID (canvas course ID), submission_type =>
  external_tool, and custom values for the rest of the fields. The
  external tool URL can be something like
  http://web.lti-13-test-tool.docker/launch?test=123
- Check that the resource_link_id field is returned in the result
- Check in a local Canvas rails console that one new assignment was
  created, that that assignment only has one line item (a.line_items), and
  that that line_item has a resource_link with the
  resource_link.resource_link_id in the response body. Also check that the
  line_item has the "tag" value you entered, and "extensions" (with the
  submission type and tool URL)
- open the assignment (which will launch the tool) and make sure the
  right resource_link_id is sent in the launch id_token, and it goes to
  the right URL (e.g http://web.lti-13-test-tool.docker/launch?test=123)
- Create another line item, this time with submission_type=none
- Check that an assignment is created with just one line_item.
- Create another line item, this time with submission_type=external_url
  again but also with a resourceLinkId equal to the resource link ID from
  before. You will have to modify the test tool
  (app/controllers/line_items_controller.rb -- create_or_update_body) to
  allow the new field and modify the HTML form.
- Check that no new assignments are created, but that a new LineItem,
  belonging to the existing Assignment, is created.

Change-Id: Ib7982c8dfdc1ea3fe3e67abf3f0c99353e9fcf23
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/236768
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Karl Lloyd <karl@instructure.com>
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>