Allow use of mobile login QR codes to be disabled
Closes USERS-559
flag=none

For some customers with MFA enabled, allowing a mobile login
via a QR code is a security issue since that authentication
will not use MFA.

Since the feature flag for QR login was forced on at the top
level, there is no way for customers to turn it off themselves.

So, let's remove the FF entirely, and instead create an account
setting (default true) for allowing mobile QR login. Then a
customer account admin can turn it off if they choose.

This commit leaves a bit of a hole in that it is still possible
to manually issue a POST request to get a QR code. But that
code is in the instructure_misc_plugin repo so I will have to
take care of that in a different commit once this one merges to
canvas_lms.

Test plan:
* You will need instructure_misc_plugin installed and have
  gone through the steps to set up an account for QR login
  (see the instructions in g/232750 if you have not already
  done this)
* Verify that the "QR for Mobile Login" feature flag is gone
* Go to account settings
* Verify that the new setting "Allow logins to the mobile apps
  via the use of QR codes" appears (it's right under the MFA
  setting since that seemed to be the most relevant place) and
  that it is checked (enabled)
* Verify that mobile QR login works as it always has

* Now uncheck that setting and "Update Settings" way at the bottom

* Now the "QR for Mobile Login" link should no longer appear in
  the account tray
* Try navigating by hand to /profile/qr_mobile_login ... that
  should now 404

Change-Id: Ic513870b00bf1b574bc00d008b4e00a31541bc8a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/237658
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Nick Houle <nhoule@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>