Treat RCE uploaded SVG files as images
There are a couple places in canvas doesn't treat uploaded svg files
as images. When returned from the files api, svg files are missing
a thumbnail url, and the response from canvas on uploading an svg file
misclassified it as a 'file' and not an 'image'

This commit reverts some generalization introduced into mime class categorization.
In an attempt to keep from having to play whack-a-mole with new mime types, I had
introduced changes where anything "image/*" was an image, and the same for "audio/*"
and "video/*", but there are a _lot_ of types I didn't consider that we wouldn't
be able to handle. For example, safari doesn't handle image/webp, so we shouldn't
yet treat that as an image file type.

closes LA-860
flag = rce_enhancements

test plan:
  - in the RCE, upload an svg file
  > expect it to be previewed in the image > upload dialog
  > expect it to be embedded (not linked) in the RCE
  - inspect the "create_success" network request's response
  > expect the returned "mime_class" to be "image"

  - from images > user or course images
  > expect the tray to display the svg
  - select the svg
  > expect it to be embedded, not linked

  - from media > record and upload, record audio
  > expect it to show up in the rce as the embedded
    audio player, and not as a link (this confirms
    a fix from an earlier ticket wasn't broken)

Change-Id: I6a2b27550c674555e4f62a97a18091467d094b71
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/239633
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>