be cool with too many access tokens in settings
fixes USERS-369

the large amount of Access Tokens was making ERB rendering take longer
than it did for the JS to be executed, while that was expecting elements
to be in the DOM that weren't.

Instead of wrapping the whole file itself (profile.js) to execute on
DOMContentLoaded or such, I opted to introduce a new API parallel to
js_bundle that abstracts this if only to preserve the history of that
file.

| TEST PLAN |
| ---- ---- |

- before you check out this patch, repro the issue by creating a ton of
  access tokens in the console and visit the settings page

    1000.times { User.first.access_tokens.create! }

- don't open your dev console, go to the settings page and hit Edit
  Settings and verify it does nothing

- check out the patch, rerun webpack and (hard) reload the page
- Edit Settings should be ok now

Change-Id: I51bb8cdb9cd91107166c6d4c0835c3709a6f10f0
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/239958
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>