Do not set RCE images to be so small
In the old RCE, images were added and let the prevailing CSS
constrain the width to 100%.  With COREFE-241, I incorrectly
thought that was poor UX and constrained the images to be 320px
on the long side when first embedded. Turns out I was wrong.
This change reverts that and lets the images be their natural size,
subject to the 100% max-width constraint, just like the old RCE

closes LA-1127
flag = rce_enhancements

test plan:
  - add a large photo to the RCE
  > expect it to be constrained in size only by the width of
    the RCE
  > expect the RCE's height to expand to the eventual rendered
    size of the image, and not the image's _real_ height
  - save
  > expect the same result
  - edit again
  - use the Image Options tray or drag handles to change
    its size
  > expect it to respect the new size
  - save
  > expect it to look like it did before

Change-Id: I9046f1fd2d9dd3af7435e75ff30bc1031e823f02
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/239976
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>