Pseudonym root_account_id *is* the account_id
Closes USERS-423
flag=none

In seeing if anything needed to be done for CD2 in the
pseudonym model, it looks like it's unnecessarily hitting
the account model to get the root account, when the account
of a pseudonym already MUST be a root account. So we get a
tiny optimization by ripping that extra table reference out.

Test plan:
* you agree

Change-Id: Icaabb9117fb1fc4f26699d9d4b1c5dba027debf9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/240230
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>