Fix up all <Spinner> import locations
Closes FOO-649
flag=none

Note to reviewers:  THIS LOOKS MUCH WORSE THAN IT IS!

Change the import location of the InstUI <Spinner>
component from the backward-compatible ui-elements to
ui-spinner. This touches a lot of files but because the
component itself works the same, nearly always this is
just a change to the imports in the module. Occasionally
I had to delete a class or fix up the deprecated "title"
prop.

This doesn't catch the stuff in packages/ ... I'll fix up
those in another commit.

Test plan:
* Just spot check some of these to make sure they still
  work? Probably isn't necessary to check every single
  one of these since the risk here is very very low.

Change-Id: I7b27590728859a69e8f44f6b3fd4e62e6e6fcf15
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/240804
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>