Change canvas_quizzes dependencies to use https:// instead of git://
The canvas_quizzes dependency for grunt-template-jasmine-requirejs used
the git protocol (`git://`). This is an issue when trying to build in a Docker
environment AND when behind a corporate firewall. Git can be configured
to prefer `https://` instead of `git://`; however, this configuration
takes place in the host's gitconfig file and isn't visible to the git
implementation inside a container (unless you mount your host's
gitconfig inside the container, which may or may not be desirable).

This commit changes canvas_quizzes' package.json file to use `https://`
instead of `git://`. This resolves the issue of not being able to fetch
the dependency from inside a container whilst behind a firewall since
Docker will pick up the `https_proxy` environment variables from the
host and use the proxy to fetch the dependency via https.

Test plan:

- check out the repo
- run `yarn`

Expected result:

- all dependencies should be fetched without issue

Change-Id: I866ae05b120c190bfe268b936fb873f5fdc98e75
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/241575
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>