allow site admins to disable GA & FullStory
fixes FOO-692
flag  = none

note that we don't support controlling these settings through the API
since it's not part of the requirements

\ test plan \
 \ --------- \

- as a site admin, verify you can toggle Google Analytics and FullStory
  for a root account other than the siteadmin account (details below)
- as an account admin, verify you do NOT see the privacy controls

=======================================================================

If you want to fully QA this, I'll walk you through enabling the two
offenders first and then checking that they're running.

== Google Analytics

To enable GA, configure the `google_analytics_key` in the console and
restart your Rails server:

    Setting.set('google_analytics_key', 'asdfasdf')

To confirm that GA is disabled, its snippet should not be included in
the DOM. You can use this query to get at that element:

    [].filter.call(
      document.head.querySelectorAll('script'),
      x => x.innerHTML.match(/window\.ga/)
    )

== FullStory

To enable FS in the first place, you must first enable the feature flag
"enable_fullstory" at the site_admin level and then configure your
dynamic_settings.yml as shown in g/230591. Be sure to restart Rails.

To confirm that FS is disabled, its snippet should not be included in
the DOM. You can use this query to get at that element:

    [].filter.call(
      document.head.querySelectorAll('script'),
      x => x.innerHTML.match(/FS\.identify/)
    )

Change-Id: Ib7493f6dc342f4d96cb0c6fa944b7e56e8a45d29
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242531
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>