Allow verbose "Excused" on gradebook upload review
On the Gradebook CSV upload confirmation page, treat values of "Excused"
as actually representing an excusal, as opposed to only values of "EX".
This was confusing because the latter would appear to be expanded in to
the former.

flag=none
fixes EVAL-53

Test plan:
- Have a course with some students and assignments
- Export a CSV
- Make some changes and re-import your CSV (or change values in the CSV)
- On the upload confirmation page, change at least one value to
  "Excused"
  - Type it out fully--don't simply enter "EX" and let it be expanded
- Save changes
- On returning to the gradebook, check that the relevant item is
  actually marked as excused

Change-Id: If5b65a289cceee72a92fefa9832daca572419031
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242705
Reviewed-by: Syed Hussain <shussain@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Adrian Packel <apackel@instructure.com>
Product-Review: Syed Hussain <shussain@instructure.com>