RCE creates links to images and media when content is selected
Inserting images or media in the RCE from the tray typically embeds it, but
if the user has selected text, the text is getting replaced with the embed,
which is unexpected. This PS changes it so if there is something
selected, we create a link to it.

Except when uploading/recording/embedding. In that case the uploaded image
or uploaded, recorded, embedded media is embedded just after the selected
text.

We also update the CanvasMediaPlayer so that it correctly handles
being not in an iframe, but window.top, as is the case when
the user clicks on the created link

closes LS-1191
flag=rce_enhancements

test plan:
  - enter some text in the RCE
  - select some text
  - go to Media toolbar button and upload or record media
  > expect the media to be embedded in the content just after
    the selected text
  - repeat by uploading an image

  - select some text
  - go to Media toolbar button > User Media
  - select a media file
  > expect the selected text to be a link to the video
  - repeat using Image > Course or User Image

  - save
  - click on the media link
  > expect a new tab to open and show the media or image.
  > expect the player to look reasonable
  > expect resizing the window to work reasonably well. Tthere
    are some edge cases not accounted for. (portrait videos, if
    shrunk vertically don't grow when the window gets larger.
    landscape videos don't shrink in height when the browser is
    made short)
  - repeat for audio

  - click on the image link
  > expect the image to be downloaded (we have another ticket
    to make file links behave more uniformly)

Change-Id: I70ab14aaa593d48688d99b0ba006f60a300ed8a6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242726
Reviewed-by: Alex Anderson <raanderson@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>