Fix error in swizzling auto-preview link's href
when a link to a file has a preview (auto-inline or not), there's
code that edits its href to remove "download" from the path and
"wrap=1" from the query_string. There was a bug that if the href
included a verifier the result was something like
"/users/2/files/17&verifier=xxx", which is invalid. This tweaks
the href editing so the result looks like
"/users/2/files/17?verifier=xxx", which is valid and works.

closes LA-1170, LA-1083
flag=none

test plan:
  - as a teacher, create a discussion
  - as a student in the course, reply and upload 2 pdfs
  - in the first, use link options to set to auto preview
  - leave the 2nd alone
  - save
  - as the teacher
    > expect the 1st file to auto-preview
    > expect the 2nd to preview if you click the icon
    > expect to be able to click on eiher file w/o error
      (wlll either preview or download)
  - as different student in the course, view the discussion
    > expect the same results as the teacher

Change-Id: I4b03be74d9536ee645e335e7945e17cd7a7ef98b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242791
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>