allow special characters in external feeds
When creating an announcement external feed in the UI with the 'Only add
posts with a specific phrase in the title' option enabled, the # and &
characters would remove parts of the string. This was due to us trying
to add the parameters as url parameters in the post rather than just
passing them in to the axios.post method as parameters.

fixes VICE-548
flag=none

/ ---- ---- \
| Test Plan |
\ ---- ---- /

- From the announcements page of a course create an external feed
- Enable the 'Only add posts with a specific phrase in the title' option
  and use the # or & characters as part of the string (e.g. #howdy)
- Navigate to `/api/v1/courses/<course_id>/external_feeds`
- Note that the feed should contain the full header match including the
  special characters

Change-Id: I476d6afcf7dcbd61bfd57b31d539aec5f88a2d21
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242795
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Davis Hyer <dhyer@instructure.com>
QA-Review: Davis Hyer <dhyer@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>