Add JavaScript to sanitize user generated CSS
On every page, a mutation observer is created for the `#content` element.
Whenever stuff is added to the page, it will run a little sanitize script.

This goes to every `.user_content` element, loops through all of the
children, and checks to see if it has `position:fixed` applied to it.
If it does, it replaces it with `position:relative`.

See the attached JIRA for an explanation of the security vulnerability
this fixes.

Test Plan:
In an RCE on any page, include the following iframe content:

```<iframe style="position: fixed;
left: 0;
top: 0;
right: 0;
width: 100vw;
height: 100vh;"
src="https://www.badgerbadgerbadger.com/"></iframe>```

Save it and navigate to the page which renders your content.
If you see badgers across the entire page, this commit failed.

Fixes SEC-2876
Refs LS-743

flags=none

Change-Id: I11bedc6946e851085503e87f892d3f9edac64bd5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/243025
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>