spec: prefer redis cache store
I decided to go this direction for two reasons. We are using
Redis for caching in production, and we want this test to mirror
that environment as closely as possible. Also it seemed the error
was timing related when doing a lookup on notifications,
memcache got deleted when the cached objects were expected to be
there. Having a separate cache store will add a more reliable
persistence layer.

fixes FOO-755
flag = none

test plan:
 - specs pass
 - no flaky flaky

Change-Id: I5db414d5ed0d47250d3247e63a81fdef0157b8ce
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/243812
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>