fix destroy batch to actually delete enrollments
this was a regression from when destroy batch was introduced, but it's
been low impact cause not many systems have deleted courses with active
enrollments. Also when a course is deleted, a user loses access to the
course anyway.
There is a spec that runs through this, but it had the wrong count on
the expectation cause there were more deleted enrollments than active
enrollments in the course in the test.

test plan
 - delete a course with active enrollments
 - in console see that enrollments were deleted
 - restore sis batch states
 - the course and enrollments should be restored

fixes VICE-741
flag=none

Change-Id: I8b755f9e827df4fa1a9276162a53aba9f25a47f1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/244308
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Matthew Lemon <mlemon@instructure.com>
QA-Review: Matthew Lemon <mlemon@instructure.com>
Product-Review: Matthew Lemon <mlemon@instructure.com>