render account np for users with no enrollments
A user with no enrollments would not be able to render the new account
notification preferences page. This change makes it so that the user is
returned in the graphql query if the user is the same as the current
user.

fixes VICE-682
flag=notification_update_account_ui

/ ---- ---- \
| Test Plan |
\ ---- ---- /

- Turn on the notificaiton_update_account_ui ff
- login as a user with no enrollments
- navigate the the account notification preferences page
- Note that the page should load correctly

Change-Id: I81b9c9daaef58c4b8a50880c1c13f6ceec2c9d3f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/244695
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Caleb Guanzon <cguanzon@instructure.com>